fix killing machine achievement by parsing the time correctly

+++ src/UltraAchievementsRevamped.Mod/Achievements/KillingMachine.cs
@@ -12,7 +12,7 @@ internal class KillingMachine
     {
         string[] parts = __instance.time.text.Split(':');
         if (parts.Length < 2) return;
-        if (!int.TryParse(parts[0], out int minutes) || !int.TryParse(parts[1], out int secs)) return;
+        if (!int.TryParse(parts[0], out int minutes) || !float.TryParse(parts[1], out float secs)) return;
 
         float seconds = minutes * 60 + secs;
         bool perfectRank = __instance.totalRank.text.Contains(">P<");
